Rename LocalizationsService to LanguagePackService and misc moves (#150314)
* rename LocalizationsService to LanguagePackService and misc moves

* couple more renames

* one more rename
