+++ src/vs/platform/driver/browser/driver.ts
@@ -7,7 +7,7 @@ import { getClientArea, getTopLeftOffset } from 'vs/base/browser/dom';
 import { coalesce } from 'vs/base/common/arrays';
 import { language, locale } from 'vs/base/common/platform';
 import { IElement, ILocaleInfo, ILocalizedStrings, IWindowDriver } from 'vs/platform/driver/common/driver';
-import localizedStrings from 'vs/platform/localizations/common/localizedStrings';
+import localizedStrings from 'vs/platform/languagePacks/common/localizedStrings';
 
 export class BrowserWindowDriver implements IWindowDriver {
