+++ src/vs/platform/languagePacks/common/localizedStrings.ts
