+++ src/vs/workbench/contrib/localization/electron-sandbox/minimalTranslations.ts
